feat(songbook): add access control and UX improvements
- Added IsGranted annotations to restrict access to Songbook routes
- Implemented security checks to ensure users can only access their own songbooks
- Improved user experience with better feedback and navigation flow

+++ src/Controller/SongbookController.php
@@ -10,8 +10,10 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 
 #[Route('/songbook')]
+#[IsGranted('ROLE_USER')]
 final class SongbookController extends AbstractController
 {
     #[Route('', name: 'app_songbook_index', methods: ['GET'])]
@@ -37,6 +39,7 @@ final class SongbookController extends AbstractController
 
             $entityManager->persist($songbook);
             $entityManager->flush();
+            $this->addFlash('success', 'Le songbook a bien été créé.');
 
             return $this->redirectToRoute('app_songbook_index', [], Response::HTTP_SEE_OTHER);
         }
@@ -50,6 +53,9 @@ final class SongbookController extends AbstractController
     #[Route('/{id}', name: 'app_songbook_show', methods: ['GET'])]
     public function show(Songbook $songbook): Response
     {
+        if ($songbook->getUser() !== $this->getUser()) {
+            throw $this->createAccessDeniedException('Vous ne pouvez pas accéder à ce songbook.');
+        }
         return $this->render('songbook/show.html.twig', [
             'songbook' => $songbook,
         ]);
@@ -58,11 +64,15 @@ final class SongbookController extends AbstractController
     #[Route('/{id}/edit', name: 'app_songbook_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Songbook $songbook, EntityManagerInterface $entityManager): Response
     {
+        if ($songbook->getUser() !== $this->getUser()) {
+            throw $this->createAccessDeniedException('Vous ne pouvez pas accéder à ce songbook.');
+        }
         $form = $this->createForm(SongbookForm::class, $songbook);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager->flush();
+            $this->addFlash('success', 'Le songbook a bien été mis à jour.');
 
             return $this->redirectToRoute('app_songbook_index', [], Response::HTTP_SEE_OTHER);
         }
@@ -76,9 +86,14 @@ final class SongbookController extends AbstractController
     #[Route('/{id}', name: 'app_songbook_delete', methods: ['POST'])]
     public function delete(Request $request, Songbook $songbook, EntityManagerInterface $entityManager): Response
     {
+        if ($songbook->getUser() !== $this->getUser()) {
+            throw $this->createAccessDeniedException('Vous ne pouvez pas accéder à ce songbook.');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$songbook->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($songbook);
             $entityManager->flush();
+            $this->addFlash('danger', 'Le songbook a bien été supprimé.');
         }
 
         return $this->redirectToRoute('app_songbook_index', [], Response::HTTP_SEE_OTHER);
